+++ src/svd/tests/test4.c
@@ -196,4 +196,19 @@ int main(int argc,char **argv)
          suffix: 4_hip_cross
          args: -svd_type cross
 
+   testset:
+      args: -svd_monitor_cancel -mat_type aijkokkos
+      requires: hip kokkos_kernels !single
+      filter: grep -v "Transpose mode" | sed -e "s/seqaijkokkos/seqaij/"
+      output_file: output/test4_1.out
+      test:
+         suffix: 4_kokkos_lanczos
+         args: -svd_type lanczos
+      test:
+         suffix: 4_kokkos_trlanczos
+         args: -svd_type trlanczos -svd_ncv 12
+      test:
+         suffix: 4_kokkos_cross
+         args: -svd_type cross
+
 TEST*/