+++ src/sys/classes/bv/tests/test11.c
@@ -234,6 +234,10 @@ int main(int argc,char **argv)
          suffix: 1_hip
          args: -bv_type {{svec mat}} -vec_type hip
          requires: hip
+      test:
+         suffix: 1_kokkos
+         args: -bv_type mat -vec_type kokkos
+         requires: hip kokkos_kernels
 
    testset:
       args: -withb -bv_orthog_block {{gs chol svqb}}
@@ -250,6 +254,10 @@ int main(int argc,char **argv)
          suffix: 4_hip
          args: -bv_type {{svec mat}} -vec_type hip -mat_type aijhipsparse
          requires: hip
+      test:
+         suffix: 4_kokkos
+         args: -bv_type mat -vec_type kokkos -mat_type aijkokkos
+         requires: hip kokkos_kernels
 
    testset:
       args: -resid -bv_orthog_block {{gs chol tsqr tsqrchol svqb}}