+++ src/sys/classes/bv/tests/test19.c
@@ -145,7 +145,7 @@ int main(int argc,char **argv)
    testset:
       nsize: {{1 2}}
       output_file: output/test19_1.out
-      filter: grep -v Process | grep -v Object | sed -e 's/mpi/seq/' | sed -e 's/seqcuda/seq/' | sed -e 's/seqaijcusparse/seqaij/' | sed -e 's/seqhip/seq/' | sed -e 's/seqaijhipsparse/seqaij/' | sed -e 's/nc=2/nc=0/'
+      filter: grep -v Process | grep -v Object | sed -e 's/mpi/seq/' -e 's/seqcuda/seq/' -e 's/seqaijcusparse/seqaij/' -e 's/seqhip/seq/' -e 's/seqaijhipsparse/seqaij/' -e 's/seqkokkos/seq/' -e 's/seqaijkokkos/seqaij/' -e 's/nc=2/nc=0/'
       test:
          suffix: 1
          args: -nc {{0 2}} -bv_type {{svec mat}}
@@ -157,5 +157,9 @@ int main(int argc,char **argv)
          suffix: 1_hip
          args: -nc {{0 2}} -bv_type {{svec mat}} -mat_type aijhipsparse
          requires: hip
+      test:
+         suffix: 1_kokkos
+         args: -nc {{0 2}} -bv_type mat -mat_type aijkokkos
+         requires: hip kokkos_kernels
 
 TEST*/